can convert a row to bytes, maybe

+++ src/db/schema.rs
@@ -14,13 +14,13 @@ pub enum Type {
 #[derive(Debug)]
 pub struct TypeDef {
     // id to be used in the on disk serialization
-    pub id: u8,
+    pub id: u16,
     dbtype: Type,
 }
 // owned by a Stream
 #[derive(Debug)]
 pub struct Schema {
-    num_fields: u8,
+    num_fields: u16,
     // maps a name to an id
     fields: HashMap<String, TypeDef>,
 }
@@ -30,7 +30,7 @@ impl Schema {
         Schema{num_fields: 0, fields: HashMap::new()}
     }
 
-    pub fn add_type(&mut self, name: &str, dbtype: Type) -> u8 {
+    pub fn add_type(&mut self, name: &str, dbtype: Type) -> u16 {
         let next_id = self.num_fields + 1;
         let typedef = TypeDef{ id: next_id, dbtype: dbtype};
         self.fields.insert(name.to_string(), typedef);
